+++ runtime/src/snapshot_utils.rs
@@ -15,6 +15,7 @@ use {
     },
     agave_fs::{
         FileInfo,
+        buffered_reader::large_file_buf_reader,
         buffered_writer::{SizeLimitedWriter, large_file_buf_writer},
         io_setup::IoSetupState,
     },
@@ -42,7 +43,7 @@ use {
         account_storage::AccountStorageMap,
         account_storage_entry::AccountStorageEntry,
         accounts_db::AtomicAccountsFileId,
-        accounts_file::{AccountsFile, StorageAccess},
+        accounts_file::AccountsFile,
         utils::{ACCOUNTS_RUN_DIR, ACCOUNTS_SNAPSHOT_DIR, move_and_async_delete_path},
     },
     solana_clock::Slot,
@@ -60,6 +61,7 @@ use {
         thread,
     },
     tempfile::TempDir,
+    wincode::io::std_read::ReadAdapter,
 };
 
 pub mod snapshot_storage_rebuilder;
@@ -71,6 +73,8 @@ pub mod snapshot_storage_rebuilder;
 pub const MAX_OBSOLETE_ACCOUNTS_FILE_SIZE: u64 = 1024 * 1024 * 1024 * 12; // 12 GB
 pub const MAX_SNAPSHOT_DATA_FILE_SIZE: u64 = 32 * 1024 * 1024 * 1024; // 32 GiB
 const MAX_SNAPSHOT_VERSION_FILE_SIZE: u64 = 8; // byte
+/// Buffer size that allows several concurrent reads using default io-uring reader read size (1MiB)
+const OBSOLETE_ACCOUNTS_READ_BUF_SIZE: usize = 4 * 1024 * 1024;
 
 // Snapshot Fastboot Version History
 // Legacy - No fastboot version file, storages flushed file presence determines if snapshot is loadable
@@ -769,7 +773,11 @@ fn deserialize_obsolete_accounts(
     let obsolete_accounts_path = bank_snapshot_dir
         .as_ref()
         .join(snapshot_paths::SNAPSHOT_OBSOLETE_ACCOUNTS_FILENAME);
-    let obsolete_accounts_file = fs::File::open(&obsolete_accounts_path)?;
+    let obsolete_accounts_reader = ReadAdapter::new(large_file_buf_reader(
+        &obsolete_accounts_path,
+        OBSOLETE_ACCOUNTS_READ_BUF_SIZE,
+        &IoSetupState::default(),
+    )?);
     // If the file is too large return error
     let obsolete_accounts_file_metadata = fs::metadata(&obsolete_accounts_path)?;
     if obsolete_accounts_file_metadata.len() > maximum_obsolete_accounts_file_size {
@@ -783,7 +791,7 @@ fn deserialize_obsolete_accounts(
     }
 
     Ok(serde_snapshot::deserialize_wincode_from(
-        obsolete_accounts_file,
+        obsolete_accounts_reader,
     )?)
 }
 
@@ -1064,7 +1072,6 @@ pub fn verify_and_unarchive_snapshots(
     full_snapshot_archive_info: &FullSnapshotArchiveInfo,
     incremental_snapshot_archive_info: Option<&IncrementalSnapshotArchiveInfo>,
     account_paths: &[PathBuf],
-    storage_access: StorageAccess,
     io_setup: &IoSetupState,
 ) -> Result<(UnarchivedSnapshots, UnarchivedSnapshotsGuard)> {
     check_are_snapshots_compatible(
@@ -1089,7 +1096,6 @@ pub fn verify_and_unarchive_snapshots(
         full_snapshot_archive_info.archive_format(),
         next_append_vec_id.clone(),
         None,
-        storage_access,
         io_setup,
     )?;
 
@@ -1117,7 +1123,6 @@ pub fn verify_and_unarchive_snapshots(
             incremental_snapshot_archive_info.archive_format(),
             next_append_vec_id.clone(),
             Some(incremental_snapshot_archive_info.base_slot()),
-            storage_access,
             io_setup,
         )?;
         (
@@ -1310,7 +1315,6 @@ fn unarchive_snapshot(
     archive_format: ArchiveFormat,
     next_append_vec_id: Arc<AtomicAccountsFileId>,
     base_slot: Option<Slot>,
-    storage_access: StorageAccess,
     io_setup: &IoSetupState,
 ) -> Result<UnarchivedSnapshot> {
     let unpack_dir = tempfile::Builder::new()
@@ -1349,7 +1353,6 @@ fn unarchive_snapshot(
                         num_rebuilder_threads,
                         next_append_vec_id,
                         SnapshotFrom::Archive,
-                        storage_access,
                         None,
                     )?,
                     measure_name
@@ -1382,7 +1385,6 @@ fn spawn_streaming_snapshot_dir_files(
     snapshot_file_path: PathBuf,
     snapshot_version_path: PathBuf,
     account_paths: &[PathBuf],
-    writable: bool,
 ) -> (Receiver<FileInfo>, thread::JoinHandle<Result<()>>) {
     let (file_sender, file_receiver) = crossbeam_channel::unbounded();
     let account_paths = account_paths.to_vec();
@@ -1399,7 +1401,7 @@ fn spawn_streaming_snapshot_dir_files(
                 for dir_entry_result in fs::read_dir(account_path)? {
                     let dir_entry = dir_entry_result?;
                     let path = dir_entry.path();
-                    let file_info = FileInfo::new_from_path_writable(path, writable)?;
+                    let file_info = FileInfo::new_from_path(path)?;
                     file_sender.send(file_info)?;
                 }
             }
@@ -1418,7 +1420,6 @@ pub(crate) fn rebuild_storages_from_snapshot_dir(
     snapshot_info: &BankSnapshotInfo,
     account_paths: &[PathBuf],
     next_append_vec_id: Arc<AtomicAccountsFileId>,
-    storage_access: StorageAccess,
 ) -> Result<(
     AccountStorageMap,
     BankFieldsToDeserialize,
@@ -1497,12 +1498,10 @@ pub(crate) fn rebuild_storages_from_snapshot_dir(
 
     let snapshot_file_path = snapshot_info.snapshot_path();
     let snapshot_version_path = bank_snapshot_dir.join(snapshot_paths::SNAPSHOT_VERSION_FILENAME);
-    #[expect(deprecated)]
     let (file_receiver, stream_files_handle) = spawn_streaming_snapshot_dir_files(
         snapshot_file_path,
         snapshot_version_path,
         account_paths,
-        storage_access == StorageAccess::Mmap,
     );
 
     let SnapshotFieldsBundle {
@@ -1523,7 +1522,6 @@ pub(crate) fn rebuild_storages_from_snapshot_dir(
         num_rebuilder_threads,
         next_append_vec_id,
         SnapshotFrom::Dir,
-        storage_access,
         obsolete_accounts,
     )?;
     stream_files_handle
@@ -1597,7 +1595,7 @@ pub fn purge_old_snapshot_archives(
     );
 
     let mut full_snapshot_archives =
-        snapshot_paths::full_snapshot_archives_iter(&full_snapshot_archives_dir)
+        snapshot_paths::full_snapshot_archives_iter(full_snapshot_archives_dir.as_ref())
             .collect::<Vec<_>>();
     full_snapshot_archives.sort_unstable();
     full_snapshot_archives.reverse();
@@ -1645,7 +1643,7 @@ pub fn purge_old_snapshot_archives(
     );
     let mut incremental_snapshot_archives_by_base_slot = HashMap::<Slot, Vec<_>>::new();
     for incremental_snapshot_archive in
-        incremental_snapshot_archives_iter(&incremental_snapshot_archives_dir)
+        incremental_snapshot_archives_iter(incremental_snapshot_archives_dir.as_ref())
     {
         incremental_snapshot_archives_by_base_slot
             .entry(incremental_snapshot_archive.base_slot())
@@ -2441,7 +2439,7 @@ mod tests {
                 NonZeroUsize::new(usize::MAX).unwrap(),
             );
             let mut full_snapshot_archives =
-                full_snapshot_archives_iter(&full_snapshot_archives_dir).collect::<Vec<_>>();
+                full_snapshot_archives_iter(full_snapshot_archives_dir.path()).collect::<Vec<_>>();
             full_snapshot_archives.sort_unstable();
             assert_eq!(
                 full_snapshot_archives.len(),
@@ -2700,7 +2698,6 @@ mod tests {
                 i as u32, // Incrementing id
                 1024,
                 AccountsFileProvider::AppendVec,
-                StorageAccess::File,
             ));
             snapshot_storages.push(storage);
         }
@@ -2744,7 +2741,6 @@ mod tests {
                 i as u32, // Incrementing id
                 1024,
                 AccountsFileProvider::AppendVec,
-                StorageAccess::File,
             ));
             snapshot_storages.push(storage);
         }
@@ -2780,7 +2776,6 @@ mod tests {
                 i as u32, // Incrementing id
                 1024,
                 AccountsFileProvider::AppendVec,
-                StorageAccess::File,
             ));
             snapshot_storages.push(storage);
         }